<추가> AnswerString 생성
