--- wardle_py/wardle.py
@@ -1,8 +1,0 @@
-
-class Wardle:
-
-    def __init__(self):
-        self.answerString = ''
-
-    def get_answerString(self):
-        return self.answerString
